Transition cache into a proper persistent data store
In the near future we are going to stop keeping old dumps around
permanently (T257782), so we need to save our "cache" in a persistent
manner so we'll have old statistics forever.

Split the code into two: one CLI extractor that processes new dumps and
saves the data and the webserver that only reads from the data files.

Change-Id: Iba284c3defcc1bcc56523949c064a6acef59d1a8